fix: address final whole-branch review findings (C1-C3, I1-I6)
C1: add oaOrder/oaName to the JOB_OFFERED contract (Java, JSON Schema,
OpenAPI, Python models) so multi-OA payloads validate instead of crashing
the Hub; group-based validation lives in JobOfferedPayload's pydantic
validator since JSON Schema alone can't express it. Adds a Java-generated
cross-language fixture the Python contract test round-trips.

C2: wire GET /executions/{id}'s `results` into the reducer via a new
hydrate action so the REST-poll fallback actually updates the table;
useExecutionChannel gains a poll-only mode (url: null) so the browser
stops reconnect-storming against a nonexistent WS endpoint.

C3: wire real screenshot capture (adb screencap) and Appium session
reset (Appium's /sessions API) into build_runner - neither requires a
live WebDriver session in the Hub process, since each spec manages its
own Appium session in its own subprocess.

I1: X-Hub-Token auth on EvidenceController; startup guard against the
default hub token in a prod profile. I2: heartbeat from a background
thread while a job runs. I3: structural timeout signal -> TIMEOUT
error category (retryable). I4: scheduled sweep for abandoned RUNNING
executions. I5: monotonic-order violations map to 409, not 500. I6:
duration_ms widened to BIGINT.

+++ backend/src/main/java/com/opshub/execution/api/ExecutionController.java
@@ -50,7 +50,7 @@ public class ExecutionController {
         }
     }
 
-    public record TestResultResponse(UUID id, UUID testCaseId, int attempt, String status, Integer durationMs, String errorCategory) {
+    public record TestResultResponse(UUID id, UUID testCaseId, int attempt, String status, Long durationMs, String errorCategory) {
         private static TestResultResponse from(ExecutionService.TestResultDto dto) {
             return new TestResultResponse(dto.id(), dto.testCaseId(), dto.attempt(), dto.status(), dto.durationMs(), dto.errorCategory());
         }
